backend/requirements.txt
# PTITCONNECT backend — cài trong thư mục backend:
#   pip install -r requirements.txt
Django>=5.1,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
channels>=4.0,<5
python-dotenv>=1.0,<2
requests>=2.31,<3
beautifulsoup4>=4.12,<5
Pillow>=10.0,<12

# Bật PostgreSQL (USE_POSTGRES=true trong .env) thì cần thêm:
# psycopg2-binary>=2.9,<3
